feat: improve json logging metadata for http requests
