+++ packages/backend/package.json
@@ -64,6 +64,7 @@
         "execa": "^5.0.0",
         "express": "^4.17.3",
         "express-session": "^1.17.2",
+        "express-winston": "^4.2.0",
         "fs-extra": "^10.0.0",
         "fuse.js": "^6.4.6",
         "graphile-worker": "^0.13.0",
@@ -72,7 +73,6 @@
         "js-yaml": "^4.1.0",
         "knex": "^2.4.2",
         "moment": "^2.29.4",
-        "morgan": "^1.10.0",
         "nanoid": "^3.1.31",
         "node-cache": "^5.1.2",
         "node-fetch": "^2.6.1",
