+++ packages/backend/src/logging/logger.ts
@@ -1,78 +1,3 @@
-import winston from 'winston';
-import { lightdashConfig } from '../config/lightdashConfig';
+import { winstonLogger } from './winston';
 
-const levels = {
-    error: 0,
-    warn: 1,
-    info: 2,
-    http: 3,
-    debug: 4,
-};
-
-const colors = {
-    error: 'red',
-    warn: 'yellow',
-    info: 'green',
-    http: 'magenta',
-    debug: 'white',
-};
-winston.addColors(colors);
-
-const formatters = {
-    plain: winston.format.combine(
-        winston.format.timestamp({ format: 'YYYY-MM-DD HH:mm:ss' }),
-        winston.format.uncolorize(),
-        winston.format.printf(
-            (info) =>
-                `${info.timestamp} [Lightdash] ${info.level}: ${info.message}`,
-        ),
-    ),
-    pretty: winston.format.combine(
-        winston.format.timestamp({ format: 'YYYY-MM-DD HH:mm:ss' }),
-        winston.format.colorize({ all: true }),
-        winston.format.printf(
-            (info) =>
-                `${info.timestamp} [Lightdash] ${info.level}: ${info.message}`,
-        ),
-    ),
-    json: winston.format.combine(
-        winston.format.timestamp({ format: 'YYYY-MM-DD HH:mm:ss' }),
-        winston.format.json(),
-    ),
-};
-
-const transports = [];
-if (lightdashConfig.logging.outputs.includes('console')) {
-    transports.push(
-        new winston.transports.Console({
-            format: formatters[
-                lightdashConfig.logging.consoleFormat ||
-                    lightdashConfig.logging.format
-            ],
-            level:
-                lightdashConfig.logging.consoleLevel ||
-                lightdashConfig.logging.level,
-        }),
-    );
-}
-if (lightdashConfig.logging.outputs.includes('file')) {
-    transports.push(
-        new winston.transports.File({
-            filename: lightdashConfig.logging.filePath,
-            format: formatters[
-                lightdashConfig.logging.fileFormat ||
-                    lightdashConfig.logging.format
-            ],
-            level:
-                lightdashConfig.logging.fileLevel ||
-                lightdashConfig.logging.level,
-        }),
-    );
-}
-
-const Logger = winston.createLogger({
-    levels,
-    transports,
-});
-
-export default Logger;
+export default winstonLogger;
